Move spawner options form validation code to separate module
